# Constrained version of Probability