--- a/stack.yaml.lock
+++ b/stack.yaml.lock
@@ -5,12 +5,40 @@
 
 packages:
 - completed:
-    hackage: probability-0.2.6@sha256:6d85d961d85fd5d1a35b90fe77510f6fcc6a8f20e8ed503219c38378de9cb3cd,2857
+    hackage: probability-0.2.7@sha256:3e9fe130af3bd75e791e1add902eb2caab170cc95373281c0a9626b01fc3b104,2869
     pantry-tree:
       size: 2711
-      sha256: 3ff942c55282f5e60db3f86ec1a0f69fd0134648838c23a990f1c72b8914602c
+      sha256: ee8953628fe301a29be9ef64ebd96f8c704969fec2b4e0b39243e6499911b767
   original:
-    hackage: probability-0.2.6@sha256:6d85d961d85fd5d1a35b90fe77510f6fcc6a8f20e8ed503219c38378de9cb3cd,2857
+    hackage: probability-0.2.7@sha256:3e9fe130af3bd75e791e1add902eb2caab170cc95373281c0a9626b01fc3b104,2869
+- completed:
+    hackage: random-1.2.0@sha256:195506fedaa7c31c1fa2a747e9b49b4a5d1f0b09dd8f1291f23a771656faeec3,6097
+    pantry-tree:
+      size: 1259
+      sha256: 77505a2919b4d1d84abc36c995046f661741251e58a592ed826b7c1aee1c044c
+  original:
+    hackage: random-1.2.0
+- completed:
+    hackage: splitmix-0.1@sha256:d50c4d0801a35be7875a040470c09863342514930c82a7d25780a6c2efc4fda9,5249
+    pantry-tree:
+      size: 1148
+      sha256: 2c884d06818f79030551aab192eee23734fd499361f4f4dcb4c51d1b1b2d786e
+  original:
+    hackage: splitmix-0.1
+- completed:
+    hackage: mwc-random-0.15.0.2@sha256:50082409799461f5cc468f62131f7e50d270156c6b285c072e699140630c4cc7,3246
+    pantry-tree:
+      size: 721
+      sha256: e1dda87b78f86857cb00a5dbd179738d200711a46200d62cc595e76ea077ddc9
+  original:
+    hackage: mwc-random-0.15.0.2
+- completed:
+    hackage: constrained-normal-1.0.2@sha256:b26919aa5245b35c27bd491212ecda8f5edf708343f4ba6231f4b2d31fd6956e,1763
+    pantry-tree:
+      size: 238
+      sha256: d0ebb3cb31b6d5e7741b323ca0a93704554a3116ef83c3375748f7c2fdf80497
+  original:
+    hackage: constrained-normal-1.0.2@sha256:b26919aa5245b35c27bd491212ecda8f5edf708343f4ba6231f4b2d31fd6956e,1763
 snapshots:
 - completed:
     size: 533053