… package.yaml
… src/Examples/Auctions/AuctionSupportFunctions.hs
… src/Examples/Auctions/ChooseReservePrice.hs
… src/Examples/Auctions/DutchAuction.hs
… src/Examples/Auctions/SequentialAuction.hs
… src/Examples/Auctions/SimultaneousBidAuction.hs
… src/Examples/Markov/Output.txt
… src/Examples/Markov/TwoStageMarkovTest.hs
… src/Examples/Staking/AndGateMarkov.hs
… stack.yaml